Rename levels (#20)
* Rename levels to wgr


* rename test files
