+++ python/glow/__init__.py
@@ -1,3 +1,3 @@
 from glow.glow import *
 from glow.functions import *
-from glow.levels import *
+from glow.wgr import *
