+++ python/glow/wgr/linear_model/functions.py
