+++ python/glow/wgr/linear_model/ridge_udfs.py
