+++ test-data/wgr/ridge-regression/blockedGT.snappy.parquet
